[android] #2462 moving MapFragment into the SDK
