+++ android/java/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/MapFragment.java
@@ -1,4 +1,4 @@
-package com.mapbox.mapboxsdk.testapp;
+package com.mapbox.mapboxsdk;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -7,9 +7,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.mapbox.mapboxsdk.utils.ApiAccess;
 import com.mapbox.mapboxsdk.views.MapView;
 
-// TODO move this to SDK
 public class MapFragment extends Fragment {
 
     //
@@ -40,7 +40,7 @@ public class MapFragment extends Fragment {
         mMap = (MapView) inflater.inflate(R.layout.fragment_main, container, false);
 
         // Set accessToken
-        mMap.setAccessToken(Util.getAccessToken(container.getContext()));
+        mMap.setAccessToken(ApiAccess.getToken(container.getContext()));
 
         // Need to pass on any saved state to the map
         mMap.onCreate(savedInstanceState);
@@ -108,6 +108,15 @@ public class MapFragment extends Fragment {
         // Need to retrieve any saved state from the map
         mMap.onSaveInstanceState(outState);
         super.onSaveInstanceState(outState);
+    }
+
+    @Override
+    public void onLowMemory() {
+        Log.v(TAG, "OnLowMemory");
+
+        // Need to pass on to view
+        mMap.onLowMemory();
+        super.onLowMemory();
     }
 
     //
