+++ android/java/MapboxGLAndroidSDK/src/main/res/layout/fragment_main.xml
